+++ src/Entity/LotteryPrize.php
@@ -14,6 +14,7 @@ use c975L\ConfigBundle\Contract\UserInterface;
 use c975L\CrowdfundingBundle\Repository\LotteryPrizeRepository;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
 
 #[ORM\Entity(repositoryClass: LotteryPrizeRepository::class)]
 #[ORM\Table(name: 'crowdfunding_lottery_prize')]
@@ -25,9 +26,11 @@ class LotteryPrize
     private ?int $id = null;
 
     #[ORM\Column(length: 100)]
+    #[Assert\NotBlank]
     private ?string $title = null;
 
     #[ORM\Column(type: Types::TEXT)]
+    #[Assert\NotBlank]
     private ?string $description = null;
 
     #[ORM\Column(type: 'smallint')]