+++ src/Form/LotteryPrizeType.php
@@ -43,7 +43,7 @@ class LotteryPrizeType extends AbstractType
             ])
             ->add('description', TextareaType::class, [
                 'label' => 'label.description',
-                'required' => false,
+                'required' => true,
                 'attr' => ['rows' => 3],
             ])
             ->add('rank', ChoiceType::class, [